requirements.txt
Django==4.1.13
djangorestframework==3.14.0
django-cors-headers==4.3.1
djongo==1.3.6
pymongo==3.12.3
python-decouple==3.8
djangorestframework-simplejwt==5.3.0
sqlparse==0.2.4
gunicorn==21.2.0
whitenoise==6.6.0
